added few things to pass the tests we will go over them

+++ SmallShell.cpp
@@ -29,14 +29,14 @@ void SmallShell::executeCommand(const char *cmd_line) {
             throw invalidArgs(splitCmd[0]);
         try {
             size_t index = 0;
-            unsigned int duration = stoul(splitCmd[1], &index); //seconds sent and makes turns it into int
+            int duration = stoi(splitCmd[1], &index); //seconds sent and makes turns it into int
             if(index < splitCmd[1].size() || duration < 0)
                 throw invalidArgs(splitCmd[0]);
         }catch(exception& e){
             throw invalidArgs(splitCmd[0]);
         }
 
-        unsigned int duration = stoul(splitCmd[1]);
+        int duration = stoi(splitCmd[1]);
         finishTime = startTime + duration;
         alarm(duration);
         size -= TIMEOUT_ARGS;
